Refactor SceneGraphProperty handling code in event side to make RegisterProperty work for all handle derived classes
- Merge CustomObject into Object
- Remove unnecessary virtuals returning nullptr
- Change GetSceneObject to always return Node for Actors
- Remove GetScenePropertyOwner as obsolete
- Fix incorrect baking behaviour in SceneGraphAnimator
- Remove ObjectImplHelper as redundant code
- Chain Scene Animatable and Input -Property getters correctly to base classes
- Remove Node& from Actor and use GetSceneObject instead, added a GetNode() helper
- Remove duplicate SceneObject pointers from Shader, Renderer, RenderTask and PanGesture
- Fix RenderTask to create its own SceneObject and not return public Handles in its construction
- Fix SceneGraph::RenderTask to observe Source and Camera nodes and remove observation from RenderTask to fix timing related bugs
- Fix RenderTaskList to keep intrusive pointers to RenderTasks rather than public Handles
- Fixed event handling to not use public handles for RenderTasks
- Fixed bug in shader where Scene objects were being leaked if SetProperty was called multiple times (added a test case)
- Fixed PanGestureDetector to acquire Scene object at construction (added a test case)
- Fix a bug in integer animation where target value was incorrect due to wrong rounding (added a test case)
- Added RegisterProperty test cases for Constrainer, Path and GestureDetector classes

+++ dali/public-api/animation/path.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2018 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -82,12 +82,12 @@ void Path::Sample( float progress, Vector3& position, Vector3& tangent ) const
 
 Vector3& Path::GetPoint( size_t index )
 {
-  return GetImplementation(*this).GetPoint(index);
+  return GetImplementation(*this).GetPoint( static_cast<uint32_t>( index ) );
 }
 
 Vector3& Path::GetControlPoint( size_t index )
 {
-  return GetImplementation(*this).GetControlPoint(index);
+  return GetImplementation(*this).GetControlPoint( static_cast<uint32_t>( index ) );
 }
 
 size_t Path::GetPointCount() const
